Move template filters and functions into tarbell.template, all tied to a blueprint
